pyproject.toml
[build-system]
build-backend = "uv_build"
requires = [ "uv-build>=0.9.18,<0.10" ]

[project]
name = "music-airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dependencies = [
  "apache-airflow==3.1.3",
  "apache-airflow-providers-fab",
  "deltalake>=1.2.1",
  "polars>=1.36.1",
  "python-dotenv>=1.2.1",
  "requests>=2.32.5",
  "streamlit>=1.52.2",
  "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
  "fastmcp>=2.14.1",
  "httpx>=0.28.1",
  "pre-commit>=4.5.1",
  "pyproject-fmt>=2.11.1",
  "pytest>=9.0.2",
  "pytest-cov>=7",
  "ruff>=0.14.10",
  "ty>=0.0.5",
]

[tool.ruff]
# Extend default rules with Airflow-specific rules
lint.extend-select = [ "AIR" ]

[tool.pytest.ini_options]
addopts = "--cov=music_airflow"
